Move settings around
