+++ pytest.ini
@@ -1,5 +1,5 @@
 # -- FILE: pytest.ini (or tox.ini)
 [pytest]
-DJANGO_SETTINGS_MODULE = config.settings.base
+DJANGO_SETTINGS_MODULE = config.django.base
 # -- recommended but optional:
 python_files = tests.py test_*.py *_tests.py
